### 2025-09-14 #8984 @editorialbot
[PRE REVIEW]: CGView.js: a JavaScript package for visualizing small genomes
**Submitting author:** @sciguy (<a href="http://orcid.org/0000-0002-0407-9807">Jason R. Grant</a>)
**Repository:** https://github.com/sciguy/cgview-js
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.8.0
