### 2025-11-14 17c6501
New 'along' option for track position
