### 2025-11-18 68b8ce4
Add fast draw for backbone with many contigs
