### 2025-11-19 dfe674e
Add pixelLengthThreshold when minArcLength is 0
