### 2025-11-19 a198267
minORFLength now excludes stop codon
