### 2025-11-21 04b7c21
Improving border
