+++ paper/paper.bib
@@ -1,146 +1,0 @@
-@article{cgview_2005, 
-  year = {2005}, 
-  rating = {0}, 
-  title = {{Circular genome visualization and exploration using CGView.}}, 
-  author = {Stothard, Paul and Wishart, David S}, 
-  journal = {Bioinformatics (Oxford, England)}, 
-  doi = {10.1093/bioinformatics/bti054}, 
-  pmid = {15479716}, 
-  url = {https://academic.oup.com/bioinformatics/article-lookup/doi/10.1093/bioinformatics/bti054}, 
-  abstract = {{UNLABELLED:CGView (Circular Genome Viewer) is a Java application and library for generating high-quality, zoomable maps of circular genomes. It converts XML or tab-delimited input into a graphical map (PNG, JPG or Scalable Vector Graphics format), complete with sequence features, labels, legends and footnotes. In addition to the default full view map, the program can generate a series of hyperlinked maps showing expanded views. The linked maps can be explored using any Web browser, allowing rapid genome browsing and facilitating data sharing. 
-  AVAILABILITY:CGView (the standalone application, library or applet), sample input, sample maps and documentation can be obtained from http://wishart.biology.ualberta.ca/cgview/
-
-  CONTACT:david.wishart@ualberta.ca.}}, 
-  pages = {537 -- 539}, 
-  number = {4}, 
-  volume = {21}, 
-  language = {English}, 
-  keywords = {}, 
-  month = {02}, 
-  local-url = {file://localhost/Users/jason/Documents/Papers%20Library/Stothard-Circular%20genome%20visualization%20and%20exploration%20using%20CGView--2005-Bioinformatics%20(Oxford,%20England)_1.pdf}
-}
-
-@article{proksee_2023, 
-  year = {2023}, 
-  title = {{Proksee: in-depth characterization and visualization of bacterial genomes}}, 
-  author = {Grant, Jason R and Enns, Eric and Marinier, Eric and Mandal, Arnab and Herman, Emily K and Chen, Chih-yu and Graham, Morag and Domselaar, Gary Van and Stothard, Paul}, 
-  journal = {Nucleic Acids Research}, 
-  issn = {0305-1048}, 
-  doi = {10.1093/nar/gkad326}, 
-  pmid = {37140037}, 
-  pmcid = {PMC10320063}, 
-  abstract = {{Proksee (https://proksee.ca) provides users with a powerful, easy-to-use, and feature-rich system for assembling, annotating, analysing, and visualizing bacterial genomes. Proksee accepts Illumina sequence reads as compressed FASTQ files or pre-assembled contigs in raw, FASTA, or GenBank format. Alternatively, users can supply a GenBank accession or a previously generated Proksee map in JSON format. Proksee then performs assembly (for raw sequence data), generates a graphical map, and provides an interface for customizing the map and launching further analysis jobs. Notable features of Proksee include unique and informative assembly metrics provided via a custom reference database of assemblies; a deeply integrated high-performance genome browser for viewing and comparing analysis results at individual base resolution (developed specifically for Proksee); an ever-growing list of embedded analysis tools whose results can be seamlessly added to the map or searched and explored in other formats; and the option to export graphical maps, analysis results, and log files for data sharing and research reproducibility. All these features are provided via a carefully designed multi-server cloud-based system that can easily scale to meet user demand and that ensures the web server is robust and responsive.}}, 
-  pages = {W484--W492}, 
-  number = {W1}, 
-  volume = {51}, 
-  local-url = {file://localhost/Users/jason/Documents/Papers%20Library/Grant-Proksee-%20in-depth%20characterization%20and%20visualization%20of%20bacterial%20genomes-2023-Nucleic%20Acids%20Research_1.pdf}
-}
-
-@article{phastest_2023, 
-  year = {2023}, 
[…]
