### 2025-11-25 bfb94b8
improving border
