### 2025-12-01 21b8752
Variable border width
