### 2025-12-01 cc6f119
Fix bookmark favorite json
