### 2025-12-22 b5eb6c5
Fix outside direct strand
