+++ src/WorkerFeatureExtraction.js
@@ -4,7 +4,7 @@
 
 /**
  * CGView.js – Interactive Circular Genome Viewer
- * Copyright © 2016–2025 Jason R. Grant
+ * Copyright © 2016–2026 Jason R. Grant
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
