### 2026-01-13 24bef29
update paper with new sections
