### 2026-01-13 20899d3
Move figure to remove blank space
