### 2026-01-13 debb912
Move figure
