### 2026-02-03 #8984 @abhishektiwari
👋 @cinnetcrash @haessar @JoshuaIszatt, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)?  The submission under consideration is "CGView.js: a JavaScript package for visualizing small genomes".

Thanks
Abhi
