### 2026-02-03 #8984 @jkanche
right up my alley, happy to review this.
