### 2026-02-04 #9930 @editorialbot
[REVIEW]: CGView.js: a JavaScript package for visualizing small genomes
**Submitting author:** @sciguy (<a href="http://orcid.org/0000-0002-0407-9807">Jason R. Grant</a>)
**Repository:** https://github.com/sciguy/cgview-js
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.8.0
