### 2026-02-17 #9930 @sciguy
Hi @abhishektiwari, it looks like the PDF compiles with the GitHub action. Is there something I need to do to fix the PDF issue?
