### 2026-03-02 #9930 @haessar
A couple of queries.

> Contribution and authorship: Has the submitting author (@sciguy) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@sciguy it might be worth adding a [CRediT](https://credit.niso.org/) style contribution section in the paper. The manuscript makes clear that Paul Stothard was involved in the development of the original CGView, but he has not made any code commits to this latest CGView.js and so would be useful to see exactly what his involvement has been.

> Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?

The latter is very clear in the issue history, but the actual code contributions are all from @sciguy. @abhishektiwari How can I answer in the affirmative to "multiple developers" when there is only one?
