### 2026-03-02 2dd3553
Add Author contributions; Add missing 'A' to figure
