+++ paper/paper.md
@@ -67,4 +67,9 @@ Generative AI (ChatGPT) was used occasionally for issue triage, small code sugge
 
 This work was funded by Genome Alberta and Genome Canada.
 
+# Author contributions
+
+Jason Grant: Conceptualization (equal); Methodology; Software; Visualization; Writing (original draft)
+Paul Stothard: Conceptualization (equal); Supervision; Writing (review and editing); Funding acquisition; Resources
+
 # References
