### 2026-03-02 8dd0698
Add missing periods to author contributions
