### 2026-03-02 #9930 @sciguy
@haessar, I've added a new "Author contributions" section to the paper. It's below for reference. I also noticed that the figure was missing the `A` label, so I added that as well.

> # Author contributions
>
> Jason Grant: Conceptualization (equal); Methodology; Software; Visualization; Writing (original draft).
> Paul Stothard: Conceptualization (equal); Supervision; Writing (review and editing); Funding acquisition; Resources.
