+++ paper/paper.md
@@ -29,6 +29,8 @@ Genome maps are routinely generated as a way of understanding or conveying the f
 
 Microbial and organellar genomics frequently require circular maps and fast navigation between scales. `CGView.js` addresses this need by providing circular and linear map layouts with nucleotide-level detail, accessible through an embeddable component suited to web applications. \autoref{fig:examples} shows examples of `CGView.js` circular and linear layouts, plus a zoomed view that displays base-pair detail.
 
+Microbial and organellar genomics frequently require circular maps and fast navigation between scales. Existing circular visualization tools such as Circos [@circos_2009] and the original Java-based CGView generate high-quality genome figures, but they produce static PNG/SVG outputs that must be regenerated to reflect changes, limiting interactive exploration. `CGView.js` addresses this need by providing circular and linear map layouts with nucleotide-level detail as an embeddable JavaScript component, combining smooth zooming and panning with dynamic, programmatic updates through an extensive API. \autoref{fig:examples} shows examples of `CGView.js` circular and linear layouts, plus a zoomed view that displays base-pair detail.
+
 # State of the field
 
 Several JavaScript-based genome browsers, including JBrowse [@jbrowse_2023], igv.js [@igvjs_2022], and pileup.js [@pileup_2016], are widely used for general genomics visualization. However, few support the circular maps that are often preferred for microbial and organellar genomes, and none provide the rapid and smooth zooming to the DNA sequence level available in `CGView.js`. `CGView.js` complements these tools by focusing on circular visualization and tight integration into web pipelines rather than operating as a standalone browser.
