+++ paper/paper.bib
@@ -100,6 +100,27 @@ pages = {gkaf360}
 }
 
 
+@article{circos_2009,
+  year = {2009}, 
+  rating = {0}, 
+  title = {{Circos: an information aesthetic for comparative genomics.}}, 
+  author = {Krzywinski, Martin and Schein, Jacqueline and Birol, Inanç and Connors, Joseph and Gascoyne, Randy and Horsman, Doug and Jones, Steven J and Marra, Marco A}, 
+  journal = {Genome Research}, 
+  issn = {1088-9051}, 
+  doi = {10.1101/gr.092759.109}, 
+  pmid = {19541911}, 
+  pmcid = {PMC2752132}, 
+  url = {http://genome.cshlp.org/cgi/doi/10.1101/gr.092759.109}, 
+  abstract = {{We created a visualization tool called Circos to facilitate the identification and analysis of similarities and differences arising from comparisons of genomes. Our tool is effective in displaying variation in genome structure and, generally, any other kind of positional relationships between genomic intervals. Such data are routinely produced by sequence alignments, hybridization arrays, genome mapping, and genotyping studies. Circos uses a circular ideogram layout to facilitate the display of relationships between pairs of positions by the use of ribbons, which encode the position, size, and orientation of related genomic elements. Circos is capable of displaying data as scatter, line, and histogram plots, heat maps, tiles, connectors, and text. Bitmap or vector images can be created from GFF-style data inputs and hierarchical configuration files, which can be easily generated by automated tools, making Circos suitable for rapid deployment in data analysis and reporting pipelines.}}, 
+  pages = {1639 -- 1645}, 
+  number = {9}, 
+  volume = {19}, 
+  language = {English}, 
+  keywords = {}, 
+  month = {09}, 
+  local-url = {file://localhost/Users/jason/Documents/Papers%20Library/Krzywinski-Circos-%20an%20information%20aesthetic%20for%20comparative%20genomics--2009-Genome%20Research.pdf}
+}
+
 @article{jbrowse_2023, 
   year = {2023}, 
   title = {{JBrowse 2: a modular genome browser with views of synteny and structural variation}}, 
