### 2026-03-04 d277c9c
Old section removal was not saved
