### 2026-03-04 642b52d
Fix spelling in README
