### 2026-03-18 abf787a
Refresh labels when feature name changes
