### 2026-03-31 26a2466
Refine canvas-related rendering updates across map components
