+++ src/Canvas.js
@@ -241,14 +241,24 @@ class Canvas {
 
   /**
    * Draws an arc or arrow on the map.
-   * @param {String} layer - Name of layer to draw element on
-   * @param {Number} start - Start position (bp) of element
-   * @param {Number} stop - Stop position (bp) of element
-   * @param {Number} centerOffset - Distance form center of map to draw element
-   * @param {Color} color - A string describing the color. {@link Color} for details.
-   * @param {Number} width - Width of element
-   * @param {String} decoration - How the element should be drawn. Values: 'arc', 'clockwise-arrow', 'counterclockwise-arrow', 'none'
-   * @param {Boolean} showShading - Should the elment be drawn with shading [Default: value from settings [showShading](Settings.html#showShading)]
+   *
+   * @param {Object} options - Drawing options
+   * @param {String} [options.layer='map'] - Name of layer to draw element on
+   * @param {Number} options.start - Start position (bp) of element
+   * @param {Number} options.stop - Stop position (bp) of element
+   * @param {Number} options.centerOffset - Distance from center of map to draw element
+   * @param {String} [options.color='#000000'] - A string describing the color. {@link Color} for details.
+   * @param {Number} [options.width=1] - Width of element
+   * @param {String} [options.decoration='arc'] - How the element should be drawn.
+   *   Values: 'arc', 'clockwise-arrow', 'counterclockwise-arrow', 'none'
+   * @param {Boolean} [options.showShading] - Should the element be drawn with shading
+   *   [Default: value from settings {@link Settings#showShading}]
+   * @param {Boolean} [options.showBorder] - Should the element be drawn with a border
+   *   [Default: value from settings {@link Settings#showBorder}]
+   * @param {Boolean} [options.fast=false] - Fast drawing mode
+   * @param {Boolean} [options.selected=false] - Is the element selected
+   * @param {Number} [options.minArcLength] - Minimum arc length in pixels
+   *   [Default: value from legend {@link Legend#defaultMinArcLength}]
    * @private
    */
   // Decoration: arc, clockwise-arrow, counterclockwise-arrow, none
@@ -281,28 +291,9 @@ class Canvas {
   // (ie the arc wiggle in the map as zooming)
   // So when the zoomFactor is large, switch to drawing lines ([path](#path) handles this).
 
-  /**
-   * Bridge method for drawElement that accepts a single options object.
-   * Intended to incrementally replace drawElement's positional arguments.
-   *
-   * @param {Object} options - Drawing options
-   * @param {String} [options.layer='map'] - Name of layer to draw element on
-   * @param {Number} options.start - Start position (bp) of element
-   * @param {Number} options.stop - Stop position (bp) of element
-   * @param {Number} options.centerOffset - Distance from center of map to draw element
-   * @param {String} [options.color='#000000'] - A string describing the color. {@link Color} for details.
-   * @param {Number} [options.width=1] - Width of element
-   * @param {String} [options.decoration='arc'] - How the element should be drawn.
-   *   Values: 'arc', 'clockwise-arrow', 'counterclockwise-arrow', 'none'
-   * @param {Boolean} [options.showShading] - Should the element be drawn with shading
-   *   [Default: value from settings {@link Settings#showShading}]
-   * @param {Number} [options.minArcLength] - Minimum arc length in pixels
-   *   [Default: value from legend {@link Legend#defaultMinArcLength}]
-   * @private
-   */
-  drawElementTest(options = {}) {
[…]
