### 2026-03-31 678cdef
Add 'showBorder' setting
