+++ src/Settings.js
@@ -40,6 +40,7 @@ import utils from './Utils';
  * [format](#format)                   | String    | The layout format of the map: circular, linear [Default: circular]
  * [backgroundColor](#backgroundColor) | String    | A string describing the background color of the map [Default: 'white']. See {@link Color} for details.
  * [showShading](#showShading)         | Boolean   | Should a shading effect be drawn on the features [Default: true]
+ * [showBorder](#showBorder)           | Boolean   | Should a border be drawn on the features [Default: true]
  * [arrowHeadLength](#arrowHeadLength) | Number    | Length of feature arrowheads as a proportion of the feature thickness. From 0 (no arrowhead) to 1 (arrowhead as long on the feature is thick) [Default: 0.3]
  * [initialMapThicknessProportion](#initialMapThicknessProportion) | Number  | Proportion of canvas size to use for drawing map tracks at a zoomFactor of 1 [Default: 0.1]
  * [maxMapThicknessProportion](#maxMapThicknessProportion) | Number  | Proportion of canvas size to use for drawing map tracks at max zoom level [Default: 0.5]
@@ -64,6 +65,7 @@ class Settings {
     this._geneticCode = utils.defaultFor(options.geneticCode, 11);
     this.arrowHeadLength = utils.defaultFor(options.arrowHeadLength, 0.3);
     this._showShading = utils.defaultFor(options.showShading, true);
+    this._showBorder = utils.defaultFor(options.showBorder, false);
     this.initialMapThicknessProportion = utils.defaultFor(options.initialMapThicknessProportion, 0.1);
     this.maxMapThicknessProportion = utils.defaultFor(options.maxMapThicknessProportion, 0.5);
     this.viewer.trigger('settings-update', {attributes: this.toJSON({includeDefaults: true})});
@@ -143,6 +145,18 @@ class Settings {
   }
 
   /**
+   * @member {Boolean} - Get or set whether features should be drawn with a border (Default: true).
+   */
+  get showBorder() {
+    return this._showBorder;
+  }
+
+  set showBorder(value) {
+    this._showBorder = value;
+    this.viewer.drawFull();
+  }
+
+  /**
    * @member {Boolean} - Get or set the initial width/thickness of the map as a
    * proportion of the canvas dimension (Circular: minDimension; Linear:
    * height). The width will grow/shrink with the zoomFactor (Default: 0.1).
@@ -177,7 +191,7 @@ class Settings {
   update(attributes) {
     this.viewer.updateRecords(this, attributes, {
       recordClass: 'Settings',
-      validKeys: ['format', 'backgroundColor', 'showShading', 'arrowHeadLength', 'geneticCode', 'initialMapThicknessProportion', 'maxMapThicknessProportion']
+      validKeys: ['format', 'backgroundColor', 'showShading', 'showBorder', 'arrowHeadLength', 'geneticCode', 'initialMapThicknessProportion', 'maxMapThicknessProportion']
     });
     this.viewer.trigger('settings-update', { attributes });
   }
@@ -191,6 +205,7 @@ class Settings {
       geneticCode: this.geneticCode,
       backgroundColor: this.backgroundColor.rgbaString,
       showShading: this.showShading,
+      showBorder: this.showBorder,
       arrowHeadLength: this.arrowHeadLength,
       initialMapThicknessProportion: this.initialMapThicknessProportion,
       maxMapThicknessProportion: this.maxMapThicknessProportion
