### 2026-03-31 25c153b
Add 'selected' attribute to Feature
