### 2026-03-31 23bb66f
Add borderColor to settings
