+++ src/Viewer.js
@@ -1635,6 +1635,7 @@ class Viewer {
    */
   invertAllColors() {
     this.settings.update({backgroundColor: this.settings.backgroundColor.invert().rgbaString});
+    this.settings.update({borderColor: this.settings.borderColor.invert().rgbaString});
 
     this.legend.invertColors();
     this.captions().each( (i, caption) => caption.invertColors() );
@@ -1652,6 +1653,7 @@ class Viewer {
    */
   invertMapColors() {
     this.settings.update({backgroundColor: this.settings.backgroundColor.invert().rgbaString});
+    this.settings.update({borderColor: this.settings.borderColor.invert().rgbaString});
 
     this.legend.invertColors(false);
     this.captions().each( (i, caption) => caption.invertColors() );
