### 2026-03-31 8912b83
Add borderThickness to settings
