### 2026-03-31 c669a75
Add showBorder/showShading to Backbone
