### 2026-04-10 stothard-group/cgview-js#22 @haessar
Add contributions guidance
I would recommend adding either a CONTRIBUTING.md or section to the README to outline the process by which users can contribute or report issues.
