### 2026-04-10 #9930 @haessar
I've added an issue https://github.com/sciguy/cgview-js/issues/22 which, when implemented, will ensure both myself and @jkanche can tick off the "Community guidelines" checkbox
