### 2026-04-10 stothard-group/cgview-js#22 @sciguy
I've added a CONTRIBUTING.md file to the docs directory
