### 2026-04-10 #9930 @sciguy
Hi @haessar and @jkanche, I've added a `CONTRIBUTING.md` file to the `paper` branch of the repo: https://github.com/sciguy/cgview-js/blob/paper/docs/CONTRIBUTING.md

Once the review is done, I will merge the paper branch into main.
