### 2026-04-14 stothard-group/cgview-js#22 @sciguy
The CONTRIBUTING.md file will be merged into main once the review is done
