### 2026-04-17 #9930 @abhishektiwari
@haessar - Re: "multiple developers" - consulted with fellow editors and EiC - not a hard requirement as long as we can see enough evidence of community or multiple people providing feedback  to project via issues ([e.g.](https://github.com/sciguy/cgview-js/issues?q=is%3Aissue%20state%3Aclosed)) or other mechanisms.

Intent here is to recognise any type of collaboration and iterative development based on feedback loop.
