+++ src/LabelPlacementNew.js
@@ -1,0 +1,206 @@
+//////////////////////////////////////////////////////////////////////////////
+// Label Placement New
+//////////////////////////////////////////////////////////////////////////////
+
+/**
+ * CGView.js – Interactive Circular Genome Viewer
+ * Copyright © 2016–2026 Jason R. Grant
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import LabelPlacementDefault from './LabelPlacementDefault';
+import Rect from './Rect';
+import utils from './Utils';
+
+/**
+ * Circular label placement that keeps labels ordered on each side of the map.
+ *
+ * Labels are placed in a monotonic stack on the left and right halves of the
+ * circle. This keeps overlaps out, greatly reduces line crossings, and only
+ * uses a two-segment line when a straight segment would cut through the map.
+ *
+ * @extends LabelPlacementDefault
+ * @private
+ */
+class LabelPlacementNew extends LabelPlacementDefault {
+
+  toString() {
+    return 'LabelPlacementNew';
+  }
+
+  get name() {
+    return 'new';
+  }
+
+  placeLabels(labels, outerOffset) {
+    this._rectOffsetWithoutLineLength = outerOffset + this._labelLineMarginInner + this._labelLineMarginOuter;
+
+    if (this.viewer.format !== 'circular') {
+      super.placeLabels(labels, outerOffset);
+      return;
+    }
+
+    const center = this._canvasCenter();
+    const rightLabels = [];
+    const leftLabels = [];
+
+    for (let i = 0, len = labels.length; i < len; i++) {
+      const label = labels[i];
+      const outerPt = this.canvas.pointForBp(label.bp, this.rectCenterOffset());
+      const rectOrigin = utils.rectOriginForAttachementPoint(outerPt, label.lineAttachment, label.width, label.height);
+      label._defaultRect = new Rect(rectOrigin.x, rectOrigin.y, label.width, label.height);
+      label._defaultOuterPt = outerPt;
+      label._newPlacementSide = (outerPt.x >= center.x) ? 'right' : 'left';
+      if (label._newPlacementSide === 'right') {
+        rightLabels.push(label);
+      } else {
+        leftLabels.push(label);
+      }
+    }
+
+    this._placeSide(rightLabels, 'right', center);
+    this._placeSide(leftLabels, 'left', center);
+
+    for (let i = 0, len = labels.length; i < len; i++) {
+      const label = labels[i];
+      label.attachementPt = label.rect.ptForClockPosition(label.lineAttachment);
+      this._updateLinePoints(label, outerOffset);
+    }
+  }
+
+  _placeSide(labels, side, center) {
+    if (labels.length === 0) {
+      return;
+    }
+
+    const centerGap = 8;
+    const upperLabels = [];
+    const lowerLabels = [];
+
+    for (let i = 0, len = labels.length; i < len; i++) {
+      const label = labels[i];
+      let x = label._defaultRect.x;
+      if (side === 'right') {
[…]
