+++ src/Annotation.js
@@ -23,6 +23,7 @@ import CGObject from './CGObject';
 import CGArray from './CGArray';
 import LabelPlacementDefault from './LabelPlacementDefault';
 import LabelPlacementAngled from './LabelPlacementAngled';
+import LabelPlacementNew from './LabelPlacementNew';
 import Font from './Font';
 import Color from './Color';
 import NCList from './NCList';
@@ -47,7 +48,7 @@ import utils from './Utils';
  * [font](#font)                    | String    | A string describing the font [Default: 'monospace, plain, 12']. See {@link Font} for details.
  * [color](#color)                  | String   | A string describing the color [Default: undefined]. If the color is undefined, the legend color for the feature will be used. See {@link Color} for details.
  * [onlyDrawFavorites](#onlyDrawFavorites) | Boolean   | Only draw labels for features that are favorited [Default: false]
- * [labelPlacement](#labelPlacement) | String   | The label placement method for positioning labels. Choices: 'default', 'angled' [Default: 'default']
+ * [labelPlacement](#labelPlacement) | String   | The label placement method for positioning labels. Choices: 'default', 'angled', 'new' [Default: 'default']
  * [visible](CGObject.html#visible) | Boolean   | Labels are visible [Default: true]
  * [meta](CGObject.html#meta)       | Object    | [Meta data](tutorial-meta.html) for Annotation
  *
@@ -182,7 +183,7 @@ class Annotation extends CGObject {
 
   /**
    * @member {LabelPlacement} - Set the label placement instance for both fast and full drawing.
-   * Value can be one of the following: 'default', 'angled', or a custom LabelPlacement class.
+   * Value can be one of the following: 'default', 'angled', 'new', or a custom LabelPlacement class.
    */
   set labelPlacement(value) {
     const labelPlacement = this._initialializeLabelPlacement(value);
@@ -192,7 +193,7 @@ class Annotation extends CGObject {
 
   /**
    * @member {LabelPlacement} - Get or set the label placement instance for fast drawing.
-   * Values for setting can be one of the following: 'default', 'angled', or a custom LabelPlacement class.
+   * Values for setting can be one of the following: 'default', 'angled', 'new', or a custom LabelPlacement class.
    */
   get labelPlacementFast() {
     return this._labelPlacementFast;
@@ -204,7 +205,7 @@ class Annotation extends CGObject {
 
   /**
    * @member {LabelPlacement} - Get or set the label placement instance for full drawing.
-   * Values for setting can be one of the following: 'default', 'angled', or a custom LabelPlacement class.
+   * Values for setting can be one of the following: 'default', 'angled', 'new', or a custom LabelPlacement class.
    */
   get labelPlacementFull() {
     return this._labelPlacementFull;
@@ -219,7 +220,8 @@ class Annotation extends CGObject {
       switch (nameOrClass) {
         case 'default': return new LabelPlacementDefault(this);
         case 'angled': return new LabelPlacementAngled(this);
-        default: throw new Error(`Label Placement name '${nameOrClass}' unknown. Use one of 'default', 'angled'`);
+        case 'new': return new LabelPlacementNew(this);
+        default: throw new Error(`Label Placement name '${nameOrClass}' unknown. Use one of 'default', 'angled', 'new'`);
       }
     } else {
       // Use provided custom LabelPlacement class
@@ -301,6 +303,7 @@ class Annotation extends CGObject {
       if (containsStart && containsStop) {
         label.bp = label.bpDefault;
         label.lineAttachment = label.lineAttachmentDefault;
+        label.linePoints = undefined;
[…]
