### 2026-06-01 4d69cc4
Add citation to readme
