+++ README.md
@@ -14,6 +14,14 @@
 CGView.js is a <strong>C</strong>ircular <strong>G</strong>enome <strong>View</strong>ing
 tool for visualizing and interacting with small genomes. 
 
+## Citing CGView.js
+
+Grant JR and Stothard P  
+CGView.js: a JavaScript package for visualizing small genomes.  
+Journal of Open Source Software, 2026, 11(122), 9930.  
+[https://doi.org/10.21105/joss.09930](https://doi.org/10.21105/joss.09930)
+
+
 ## Resources
 
 - [CGView.js Home](http://js.cgview.ca)
