### 2026-07-08 4933f3f
Add feature selection interactions
