### 2026-07-08 5279b54
Unselect when clicking non-features
