+++ src/Selection.js
@@ -20,6 +20,7 @@
  */
 
 import utils from './Utils';
+import CGRange from './CGRange';
 
 /**
  * The Selection object controls feature selection from viewer interactions.
@@ -41,6 +42,8 @@ class Selection {
   constructor(viewer, options = {}) {
     this.viewer = viewer;
     this._handleKeydown = this.handleKeydown.bind(this);
+    this._marquee = undefined;
+    this._suppressClick = false;
     this.enabled = utils.defaultFor(options.enabled, true);
   }
 
@@ -74,7 +77,10 @@ class Selection {
    * Enable viewer selection interactions.
    */
   attach() {
-    this.viewer.off('click.cgv-selection');
+    this.viewer.off('.cgv-selection');
+    this.viewer.on('mousedown.cgv-selection', (event) => this.handleMousedown(event));
+    this.viewer.on('mousemove.cgv-selection', (event) => this.handleMousemove(event));
+    this.viewer.on('mouseup.cgv-selection', (event) => this.handleMouseup(event));
     this.viewer.on('click.cgv-selection', (event) => this.handleClick(event));
     document.removeEventListener('keydown', this._handleKeydown);
     document.addEventListener('keydown', this._handleKeydown);
@@ -84,7 +90,9 @@ class Selection {
    * Disable viewer selection interactions.
    */
   detach() {
-    this.viewer.off('click.cgv-selection');
+    this.viewer.off('.cgv-selection');
+    this._marquee = undefined;
+    this._suppressClick = false;
     document.removeEventListener('keydown', this._handleKeydown);
   }
 
@@ -107,6 +115,10 @@ class Selection {
    * @param {Object} event - Event-like object from EventMonitor.
    */
   handleClick(event = {}) {
+    if (this._suppressClick) {
+      this._suppressClick = false;
+      return;
+    }
     const feature = this.featureFromEvent(event);
     if (!feature) {
       this.clear();
@@ -114,6 +126,54 @@ class Selection {
     }
     const append = Boolean(event.d3 && event.d3.shiftKey);
     this.select(feature, { append });
+  }
+
+  /**
+   * Start marquee selection from empty map space while Shift is pressed.
+   * @param {Object} event - Event-like object from EventMonitor.
+   */
+  handleMousedown(event = {}) {
+    const shiftKeyDown = Boolean(event.d3 && event.d3.shiftKey);
+    if (!shiftKeyDown || event.elementType) { return; }
+    this._marquee = {
+      startBp: event.bp,
+      stopBp: event.bp,
+      dragged: false,
+      initialSelectedFeatures: new Set(this.selectedFeatures()),
+      selectedFeatures: new Set()
+    };
+    if (event.d3.preventDefault) {
+      event.d3.preventDefault();
+    }
+  }
+
+  /**
+   * Update marquee selection while dragging.
+   * @param {Object} event - Event-like object from EventMonitor.
+   */
+  handleMousemove(event = {}) {
+    if (!this._marquee) { return; }
+    this._marquee.stopBp = event.bp;
+    this._marquee.dragged = this._marquee.dragged || (event.bp !== this._marquee.startBp);
+    this.selectFeaturesInMarquee();
+    this.drawMarquee();
+  }
+
+  /**
+   * Finish marquee selection and clear it from the UI layer.
+   * @param {Object} event - Event-like object from EventMonitor.
+   */
+  handleMouseup(event = {}) {
+    if (!this._marquee) { return; }
+    if (event.bp !== undefined) {
+      this._marquee.stopBp = event.bp;
+    }
+    if (this._marquee.dragged) {
+      this.selectFeaturesInMarquee();
+      this._suppressClick = true;
+    }
[…]
