### 2026-07-09 6150edd
Add Selection option to test page
