### 2026-07-09 1f7dfa3
Improve counter-clockwise marquee selection
