### 2026-07-09 2e34f75
Increase bounds of selection marquee
